backend/requirements.txt
fastapi
uvicorn
pandas
numpy==1.26.4
requests
python-dotenv
pymysql
DBUtils
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
chromadb
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.45.0
sentence-transformers==3.0.1
accelerate